dosage/requirements.txt
asgiref==3.8.1
attrs==23.2.0
black==24.4.2
certifi==2024.6.2
click==8.1.7
Django==5.1a1
django-debug-toolbar==4.4.2
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.2rc1
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.4.4
uritemplate==4.1.1
